+++ test/test_networkclustering_module.py
@@ -1,16 +1,13 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Mon Jan 31 18:11:09 2022.
-
-@author: fabian
+Tests continued functioning of deprecated `networkclustering` module.
 """
 
 import numpy as np
 import pandas as pd
 
-from pypsa.clustering.spatial import (
-    busmap_by_hac,
+from pypsa.networkclustering import (
     busmap_by_kmeans,
     get_clustering_from_busmap,
 )
@@ -22,16 +19,6 @@ def test_default_clustering_k_means(scipy_network):
     n.lines["type"] = np.nan
     weighting = pd.Series(1, n.buses.index)
     busmap = busmap_by_kmeans(n, bus_weightings=weighting, n_clusters=50)
-    C = get_clustering_from_busmap(n, busmap)
-    nc = C.network
-    assert len(nc.buses) == 50
-
-
-def test_default_clustering_hac(scipy_network):
-    n = scipy_network
-    # delete the 'type' specifications to make this example easier
-    n.lines["type"] = np.nan
-    busmap = busmap_by_hac(n, n_clusters=50)
     C = get_clustering_from_busmap(n, busmap)
     nc = C.network
     assert len(nc.buses) == 50