+++ doc/release_notes.rst
@@ -46,6 +46,10 @@ Upcoming Release
 
 * A new type of ``GlobalConstraint`` called `operational_limit` is now supported through the `Network.optimize` function. It allows to limit the total production of a carrier analogous to `primary_energy_limit` with the difference that it applies directly to the production of a carrier rather than to an attribute of the primary energy use.
 
+* The module ``pypsa.networkclustering`` was moved to
+  ``pypsa.clustering.spatial``. The module ``pypsa.networkclustering`` is now
+  deprecated but all functionality will continue to be accessible until the next
+  major version.
 
 
 PyPSA 0.22.1 (15th February 2023)